+++ src/visualize.py
@@ -37,6 +37,34 @@ class Visualizer:
         plt.axis('off')
         plt.show()
 
+    @classmethod
+    def display_rois(cls, image, feature_maps, rois):   # TODO: There appears to be a shift between the image and feature map plots
+        fig, axes = plt.subplots(1, 2, figsize=(10, 5))
+
+        image = cls._convert_to_original_image(image)
+
+        axes[0].imshow(image)
+        axes[1].imshow(np.mean(feature_maps, axis=-1)[0], cmap='gray')
+
+        scale_x = image.shape[0] / feature_maps.shape[1]
+        scale_y = image.shape[1] / feature_maps.shape[2]
+
+        for i, roi in enumerate(rois):
+            x1, y1, x2, y2 = roi
+            color = DEFAULT_COLORS[i % len(DEFAULT_COLORS)]
+
+            # Scale the coordinates to the image shape
+            ix1, iy1, ix2, iy2 = x1 * scale_x, y1 * scale_y, x2 * scale_x, y2 * scale_y
+
+            axes[0].add_patch(plt.Rectangle((ix1, iy1), ix2 - ix1, iy2 - iy1, fill=False, edgecolor=color, lw=1))
+            axes[1].add_patch(plt.Rectangle((x1, y1), x2 - x1, y2 - y1, fill=False, edgecolor=color, lw=1))
+
+        # Remove axis labels
+        for ax in axes:
+            ax.axis('off')
+
+        plt.show()
+
     @staticmethod
     def _convert_to_original_image(image):
         '''Convert the image from BGR to RGB and add the mean pixel value back to each pixel.